rmm: Expose start() to plat/fvp
